server/ml-training/requirements.txt
psycopg2-binary==2.9.9
torch==2.0.1
numpy==1.24.3
schedule==1.2.0
redis==5.0.1
watchfiles==0.21.0
